feat(AzureVpcDnsLink): rename API
